Revert "G grid layout generator"
